+++ src/design-system/vue/components/index.ts
@@ -1,1 +1,3 @@
 export { default as DsButton } from './DsButton.vue'
+export { default as DsCard } from './DsCard.vue'
+export { default as DsInput } from './DsInput.vue'